deepseek-engineer/hotppl-website/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
google-cloud-datastore==2.19.0
